transport: Trigger transport reacquire for multiple BISes sync
Once the transport is release, bap_stop_complete is called. If the
transport is part of a sync with multiple BISes, it needs to be
reacquired. Such a transport will have it's stream in the
BT_BAP_STREAM_STATE_PENDING state. Thus, transport_update_playing
with playing set to TRUE must be called. This will set the transport
state to BROADCASTING, so that the audio server will automatically
acquire it.