monitor: Allow caching of temporary gatt_db
This fixes the problem of not being able to decode GATT traffic when
reading the traces from a file/offline, the gatt_db would be discarded
upon the first disconnection so it didn't work if the trace contains
multiple connections.

To fix this btmon would remember gatt_db discovered at runtime so even
when there multiple connections the same db would be used.